Go to readyState DONE when there's a connection error.
This brings xhr.coretcpsocket in line with browser behavior.
It's also necessary for socket.io, which uses this state
change to detect errors early (in engine.io-client's polling-xhr.js)
instead of waiting and timing out.

This change also includes some reduction of extremely verbose
logging, and an additional test for TLS connection failures.